requirements.txt
Flask==3.1.2
beautifulsoup4==4.14.3
httpx==0.28.1
openpyxl==3.1.5
pandas
python-dotenv==1.1.0
python-telegram-bot==22.5
requests
